Revert "Migrate op to new infra: all_gather_async (#34184)"
This reverts commit 799fb4dbc956ddf6eff7c3f49bdf7f5a13d7444a.
