+++ ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_async/all_gather_async.cpp
@@ -4,7 +4,7 @@
 
 #include "all_gather_async.hpp"
 #include <utility>
-#include "ttnn/operations/experimental/ccl/all_gather_async/device/all_gather_async_device_operation.hpp"
+#include "ttnn/operations/experimental/ccl/all_gather_async/device/all_gather_async_op.hpp"
 #include "ttnn/operations/core/core.hpp"
 #include "ttnn/operations/data_movement/concat/concat.hpp"
 #include "ttnn/operations/copy/typecast/typecast.hpp"
@@ -25,7 +25,7 @@ ttnn::Tensor ExecuteAllGatherAsync::invoke(
     const std::optional<GlobalSemaphore>& barrier_semaphore,
     bool reverse_order,
     const std::optional<CoreRangeSet>& sub_core_grid) {
-    tt::tt_fabric::Topology usable_topology = ::ttnn::ccl::get_usable_topology(input_tensor, topology, std::nullopt);
+    tt::tt_fabric::Topology topology_ = ::ttnn::ccl::get_usable_topology(input_tensor, topology, std::nullopt);
     bool composite_all_gather_case = composite_common::use_composite_all_gather(input_tensor, dim, memory_config);
     bool all_gather_async_llama_sharded_case =
         composite_common::use_all_gather_async_llama_sharded(input_tensor, memory_config.value_or(input_tensor.memory_config()));
@@ -41,25 +41,19 @@ ttnn::Tensor ExecuteAllGatherAsync::invoke(
             /*cluster_axis*/ std::nullopt);
     } else {
         log_debug(tt::LogOp, "Using minimal_all_gather_async");
-        return ttnn::prim::all_gather_async(
+        return ttnn::operations::experimental::ccl::all_gather_async(
             input_tensor,
-            /*persistent_output_buffer*/ std::nullopt,
             dim,
             multi_device_global_semaphore,
             num_links,
             memory_config,
-            usable_topology,
+            topology_,
             subdevice_id,
-            /*cluster_axis*/ std::nullopt,
             use_optimal_ccl_for_llama,
             all_gather_async_llama_sharded_case,
             barrier_semaphore,
-            /*chunks_per_sync*/ std::nullopt,
-            /*num_workers_per_link*/ std::nullopt,
-            /*num_buffers_per_channel*/ std::nullopt,
             reverse_order,
-            sub_core_grid,
-            /*mesh_device*/ nullptr);
+            sub_core_grid);
     }
 }
 
@@ -80,7 +74,7 @@ ttnn::Tensor ExecuteAllGatherAsync::invoke(
     std::optional<uint32_t> num_buffers_per_channel,
     bool reverse_order,
     const std::optional<CoreRangeSet>& sub_core_grid) {
-    tt::tt_fabric::Topology usable_topology = ::ttnn::ccl::get_usable_topology(input_tensor, topology, cluster_axis);
+    tt::tt_fabric::Topology topology_ = ::ttnn::ccl::get_usable_topology(input_tensor, topology, cluster_axis);
     bool composite_all_gather_case = composite_common::use_composite_all_gather(input_tensor, dim, memory_config);
     bool all_gather_async_llama_sharded_case =
         composite_common::use_all_gather_async_llama_sharded(input_tensor, memory_config.value_or(input_tensor.memory_config()));
@@ -91,14 +85,14 @@ ttnn::Tensor ExecuteAllGatherAsync::invoke(
             input_tensor, dim, num_links, memory_config, subdevice_id, cluster_axis);
     } else {
         log_debug(tt::LogOp, "Using minimal_all_gather_async");
-        return ttnn::prim::all_gather_async(
+        return ttnn::operations::experimental::ccl::all_gather_async(
             input_tensor,
             persistent_output_buffer,
             dim,
             multi_device_global_semaphore,
             num_links,
             memory_config,
-            usable_topology,
+            topology_,
             subdevice_id,
             cluster_axis,
             use_optimal_ccl_for_llama,
@@ -108,8 +102,7 @@ ttnn::Tensor ExecuteAllGatherAsync::invoke(
             num_workers_per_link,
             num_buffers_per_channel,
             reverse_order,
-            sub_core_grid,
-            /*mesh_device*/ nullptr);
+            sub_core_grid);
     }
 }
 
@@ -129,8 +122,7 @@ std::vector<ttnn::Tensor> ExecuteAllGatherAsync::invoke(
     std::optional<uint32_t> num_workers_per_link,
     std::optional<uint32_t> num_buffers_per_channel,
     const std::optional<CoreRangeSet>& sub_core_grid) {
-    tt::tt_fabric::Topology usable_topology =
-        ::ttnn::ccl::get_usable_topology(input_tensors.at(0), topology, cluster_axis);
+    tt::tt_fabric::Topology topology_ = ::ttnn::ccl::get_usable_topology(input_tensors.at(0), topology, cluster_axis);
     bool composite_all_gather_case =
         composite_common::use_composite_all_gather(input_tensors.at(0), dim, memory_config);
     bool all_gather_async_llama_sharded_case = composite_common::use_all_gather_async_llama_sharded(
@@ -142,35 +134,24 @@ std::vector<ttnn::Tensor> ExecuteAllGatherAsync::invoke(
             input_tensors, dim, num_links, memory_config, subdevice_id, cluster_axis);
     } else {
         log_debug(tt::LogOp, "Using minimal_all_gather_async");
-        std::vector<Tensor> output_tensors;
-        output_tensors.reserve(input_tensors.size());
-        for (size_t i = 0; i < input_tensors.size(); ++i) {
-            // 0 = forward link, 1 = backward link, i = device index
-            std::vector global_semaphores = {
-                multi_device_global_semaphore.at(0).global_semaphores.at(i),
-                multi_device_global_semaphore.at(1).global_semaphores.at(i)};
-            output_tensors.push_back(ttnn::prim::all_gather_async(
-                input_tensors[i],
-                persistent_output_buffer,
-                dim,
-                global_semaphores,
-                num_links,
-                memory_config,
-                usable_topology,
-                subdevice_id,
-                cluster_axis,
-                use_optimal_ccl_for_llama,
-                all_gather_async_llama_sharded_case,
-                barrier_semaphore.has_value() ? std::optional<GlobalSemaphore>(barrier_semaphore.value()[i])
-                                              : std::nullopt,
-                chunks_per_sync,
-                num_workers_per_link,
-                num_buffers_per_channel,
-                /*reverse_order*/ false,
-                sub_core_grid,
-                /*mesh_device*/ nullptr));
-        }
-        return output_tensors;
+        return ttnn::operations::experimental::ccl::all_gather_async(
+            input_tensors,
+            persistent_output_buffer,
+            dim,
+            multi_device_global_semaphore,
+            num_links,
+            memory_config,
+            topology_,
+            subdevice_id,
+            cluster_axis,
+            all_gather_async_llama_sharded_case,
+            use_optimal_ccl_for_llama,
+            barrier_semaphore,
+            chunks_per_sync,
+            num_workers_per_link,
+            num_buffers_per_channel,
+            false,
+            sub_core_grid);
     }
 }
 
@@ -189,7 +170,7 @@ ttnn::Tensor ExecuteAllGatherAsync::invoke(
     const std::optional<GlobalSemaphore>& barrier_semaphore,
     bool reverse_order,
     const std::optional<CoreRangeSet>& sub_core_grid) {
-    tt::tt_fabric::Topology usable_topology = ::ttnn::ccl::get_usable_topology(input_tensor, topology, cluster_axis);
+    tt::tt_fabric::Topology topology_ = ::ttnn::ccl::get_usable_topology(input_tensor, topology, cluster_axis);
     bool composite_all_gather_case = composite_common::use_composite_all_gather(input_tensor, dim, memory_config);
     bool all_gather_async_llama_sharded_case =
         composite_common::use_all_gather_async_llama_sharded(input_tensor, memory_config.value_or(input_tensor.memory_config()));
@@ -200,25 +181,22 @@ ttnn::Tensor ExecuteAllGatherAsync::invoke(
             input_tensor, dim, num_preferred_links.value_or(1), memory_config, subdevice_id, cluster_axis);
     } else {
         log_debug(tt::LogOp, "Using minimal_all_gather_async");
-        return ttnn::prim::all_gather_async(
+        return ttnn::operations::experimental::ccl::all_gather_async(
             input_tensor,
-            persistent_output_tensor,
             dim,
-            multi_device_global_semaphore,
-            num_preferred_links.has_value() ? num_preferred_links.value() : 1,
-            memory_config,
-            usable_topology,
-            subdevice_id,
             cluster_axis,
-            use_optimal_ccl_for_llama,
+            mesh_device,
+            topology_,
+            multi_device_global_semaphore,
+            persistent_output_tensor,
+            memory_config,
+            num_preferred_links,
+            subdevice_id,
             all_gather_async_llama_sharded_case,
+            use_optimal_ccl_for_llama,
             barrier_semaphore,
-            /*chunks_per_sync*/ std::nullopt,
-            /*num_workers_per_link*/ std::nullopt,
-            /*num_buffers_per_channel*/ std::nullopt,
             reverse_order,
-            sub_core_grid,
-            &mesh_device);
+            sub_core_grid);
     }
 }
 
@@ -235,7 +213,7 @@ ttnn::Tensor ExecuteAllGatherAsyncReversed::invoke(
     bool reverse_order,
     const std::optional<CoreRangeSet>& sub_core_grid) {
     // NOTE: reverse_order parameter is ignored, always use true for reversed API
-    tt::tt_fabric::Topology usable_topology = ::ttnn::ccl::get_usable_topology(input_tensor, topology, std::nullopt);
+    tt::tt_fabric::Topology topology_ = ::ttnn::ccl::get_usable_topology(input_tensor, topology, std::nullopt);
     bool composite_all_gather_case = composite_common::use_composite_all_gather(input_tensor, dim, memory_config);
     bool all_gather_async_llama_sharded_case =
         composite_common::use_all_gather_async_llama_sharded(input_tensor, memory_config.value_or(input_tensor.memory_config()));
@@ -251,25 +229,19 @@ ttnn::Tensor ExecuteAllGatherAsyncReversed::invoke(
             /*cluster_axis*/ std::nullopt);
     } else {
         log_debug(tt::LogOp, "Using minimal_all_gather_async");
-        return ttnn::prim::all_gather_async(
+        return ttnn::operations::experimental::ccl::all_gather_async(
             input_tensor,
-            /*persistent_output_buffer*/ std::nullopt,
             dim,
             multi_device_global_semaphore,
             num_links,
             memory_config,
-            usable_topology,
+            topology_,
             subdevice_id,
-            /*cluster_axis*/ std::nullopt,
             use_optimal_ccl_for_llama,
             all_gather_async_llama_sharded_case,
             barrier_semaphore,
-            /*chunks_per_sync*/ std::nullopt,
-            /*num_workers_per_link*/ std::nullopt,
-            /*num_buffers_per_channel*/ std::nullopt,
-            /*reverse_order*/ true,
-            sub_core_grid,
-            /*mesh_device*/ nullptr);  // reverse_order=true for reversed API
+            true,
+            sub_core_grid);  // reverse_order=true for reversed API
     }
 }
 
@@ -290,7 +262,7 @@ ttnn::Tensor ExecuteAllGatherAsyncReversed::invoke(
     std::optional<uint32_t> num_buffers_per_channel,
     bool reverse_order,
     const std::optional<CoreRangeSet>& sub_core_grid) {
-    tt::tt_fabric::Topology usable_topology = ::ttnn::ccl::get_usable_topology(input_tensor, topology, cluster_axis);
+    tt::tt_fabric::Topology topology_ = ::ttnn::ccl::get_usable_topology(input_tensor, topology, cluster_axis);
     bool composite_all_gather_case = composite_common::use_composite_all_gather(input_tensor, dim, memory_config);
     bool all_gather_async_llama_sharded_case =
         composite_common::use_all_gather_async_llama_sharded(input_tensor, memory_config.value_or(input_tensor.memory_config()));
@@ -300,14 +272,14 @@ ttnn::Tensor ExecuteAllGatherAsyncReversed::invoke(
         return composite_common::composite_all_gather(input_tensor, dim, num_links, memory_config, subdevice_id, cluster_axis);
     } else {
         log_debug(tt::LogOp, "Using minimal_all_gather_async");
-        return ttnn::prim::all_gather_async(
+        return ttnn::operations::experimental::ccl::all_gather_async(
             input_tensor,
             persistent_output_buffer,
             dim,
             multi_device_global_semaphore,
             num_links,
             memory_config,
-            usable_topology,
+            topology_,
             subdevice_id,
             cluster_axis,
             use_optimal_ccl_for_llama,
@@ -316,9 +288,8 @@ ttnn::Tensor ExecuteAllGatherAsyncReversed::invoke(
             chunks_per_sync,
             num_workers_per_link,
             num_buffers_per_channel,
-            /*reverse_order*/ true,
-            sub_core_grid,
-            /*mesh_device*/ nullptr);  // reverse_order=true for reversed API
+            true,
+            sub_core_grid);  // reverse_order=true for reversed API
     }
 }
 
@@ -337,7 +308,7 @@ ttnn::Tensor ExecuteAllGatherAsyncReversed::invoke(
     const std::optional<GlobalSemaphore>& barrier_semaphore,
     bool reverse_order,
     const std::optional<CoreRangeSet>& sub_core_grid) {
-    tt::tt_fabric::Topology usable_topology = ::ttnn::ccl::get_usable_topology(input_tensor, topology, cluster_axis);
+    tt::tt_fabric::Topology topology_ = ::ttnn::ccl::get_usable_topology(input_tensor, topology, cluster_axis);
     bool composite_all_gather_case = composite_common::use_composite_all_gather(input_tensor, dim, memory_config);
     bool all_gather_async_llama_sharded_case =
         composite_common::use_all_gather_async_llama_sharded(input_tensor, memory_config.value_or(input_tensor.memory_config()));
@@ -348,25 +319,22 @@ ttnn::Tensor ExecuteAllGatherAsyncReversed::invoke(
             input_tensor, dim, num_preferred_links.value_or(1), memory_config, subdevice_id, cluster_axis);
     } else {
         log_debug(tt::LogOp, "Using minimal_all_gather_async");
-        return ttnn::prim::all_gather_async(
+        return ttnn::operations::experimental::ccl::all_gather_async(
             input_tensor,
-            persistent_output_tensor,
             dim,
-            multi_device_global_semaphore,
-            num_preferred_links.has_value() ? num_preferred_links.value() : 1,
-            memory_config,
-            usable_topology,
-            subdevice_id,
             cluster_axis,
-            use_optimal_ccl_for_llama,
+            mesh_device,
+            topology_,
+            multi_device_global_semaphore,
+            persistent_output_tensor,
+            memory_config,
+            num_preferred_links,
+            subdevice_id,
             all_gather_async_llama_sharded_case,
+            use_optimal_ccl_for_llama,
             barrier_semaphore,
-            /*chunks_per_sync*/ std::nullopt,
-            /*num_workers_per_link*/ std::nullopt,
-            /*num_buffers_per_channel*/ std::nullopt,
-            /*reverse_order*/ true,
-            sub_core_grid,
-            &mesh_device);  // reverse_order=true for reversed API
+            true,
+            sub_core_grid);  // reverse_order=true for reversed API
     }
 }
