--- ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_async/device/all_gather_async_llama_sharded_program_factory.hpp
@@ -1,42 +1,0 @@
-// SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
-//
-// SPDX-License-Identifier: Apache-2.0
-
-#pragma once
-
-#include "all_gather_async_device_operation_types.hpp"
-#include "ttnn/device_operation.hpp"
-
-namespace ttnn::operations::experimental::ccl::all_gather_async {
-
-struct LlamaShardedMeshWorkloadFactory {
-    struct shared_variables_t {
-        tt::tt_metal::KernelHandle worker_sender_reader_kernel_id;
-        tt::tt_metal::KernelHandle worker_sender_writer_kernel_id;
-        std::vector<tt::tt_metal::CoreCoord> sender_worker_cores;
-    };
-    using cached_mesh_workload_t = ttnn::device_operation::AdaptedCachedMeshWorkload<shared_variables_t>;
-
-    static cached_mesh_workload_t create_mesh_workload(
-        const operation_attributes_t& operation_attributes,
-        const ttnn::MeshCoordinateRangeSet& tensor_coords,
-        const tensor_args_t& tensor_args,
-        tensor_return_value_t& tensor_return_value);
-
-    static void override_runtime_arguments(
-        cached_mesh_workload_t& cached_workload,
-        const operation_attributes_t& operation_attributes,
-        const tensor_args_t& tensor_args,
-        tensor_return_value_t& tensor_return_value);
-
-private:
-    using cached_program_t = ttnn::device_operation::CachedProgram<shared_variables_t>;
-
-    static cached_program_t create_at(
-        const operation_attributes_t& operation_attributes,
-        const ttnn::MeshCoordinate& mesh_coordinate,
-        const tensor_args_t& tensor_args,
-        tensor_return_value_t& tensor_return_value);
-};
-
-}  // namespace ttnn::operations::experimental::ccl::all_gather_async
