+++ ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_async/device/all_gather_async_op.hpp
@@ -1,0 +1,318 @@
+// SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include <cstdint>
+#include <tt-metalium/core_coord.hpp>
+#include <tt-metalium/buffer.hpp>
+#include "ttnn/tensor/tensor.hpp"
+#include "ttnn/operations/ccl/shared_with_host/hetergeneous_data_structs.hpp"
+#include "ttnn/operations/ccl/ccl_host_datastructures.hpp"
+#include "ttnn/operations/ccl/ccl_common.hpp"
+#include "ttnn/operations/ccl/ccl_op_fusion.hpp"
+#include <tt-metalium/global_semaphore.hpp>
+#include "ttnn/global_semaphore.hpp"
+
+#include "ttnn/run_operation.hpp"
+
+#include <optional>
+#include <utility>
+#include <vector>
+
+namespace ttnn {
+
+using ccl::EriscDatamoverBuilder;
+
+enum class AllGatherAsyncVersion {
+    LLAMA_MINIMAL_SHARDED = 0,
+    MINIMAL_DEFAULT = 1,
+};
+
+struct AllGatherAsync {
+    const uint32_t dim;
+    const uint32_t num_links;
+    const uint32_t ring_size;
+    const MemoryConfig output_mem_config;
+    const ccl::Topology topology;
+    const std::vector<GlobalSemaphore> semaphore;
+    std::optional<tt::tt_metal::SubDeviceId> sub_device_id;
+    std::optional<uint32_t> cluster_axis;
+    bool use_all_gather_async_llama_sharded;
+    bool use_optimal_ccl_for_llama;
+    const std::optional<GlobalSemaphore>& barrier_semaphore;
+    bool using_persistent_buffers;
+    std::optional<uint32_t> chunks_per_sync;
+    std::optional<uint32_t> num_workers_per_link;
+    std::optional<uint32_t> num_buffers_per_channel;
+    bool reverse_order;
+    const std::optional<CoreRangeSet> sub_core_grid;
+
+    AllGatherAsync(
+        uint32_t dim,
+        uint32_t num_links,
+        uint32_t ring_size,
+        MemoryConfig output_mem_config,
+        ccl::Topology topology,
+        std::vector<GlobalSemaphore> semaphore,
+        std::optional<tt::tt_metal::SubDeviceId>& sub_device_id,
+        std::optional<uint32_t> cluster_axis,
+        bool use_all_gather_async_llama_sharded,
+        bool use_optimal_ccl_for_llama,
+        const std::optional<GlobalSemaphore>& barrier_semaphore,
+        bool using_persistent_buffers,
+        std::optional<uint32_t> chunks_per_sync,
+        std::optional<uint32_t> num_workers_per_link,
+        std::optional<uint32_t> num_buffers_per_channel,
+        bool reverse_order = false,
+        const std::optional<CoreRangeSet>& sub_core_grid = std::nullopt) :
+        dim(dim),
+        num_links(num_links),
+        ring_size(ring_size),
+        output_mem_config(std::move(output_mem_config)),
+        topology(topology),
+        semaphore(std::move(semaphore)),
+        sub_device_id(sub_device_id),
+        cluster_axis(cluster_axis),
+        use_all_gather_async_llama_sharded(use_all_gather_async_llama_sharded),
+        use_optimal_ccl_for_llama(use_optimal_ccl_for_llama),
+        barrier_semaphore(barrier_semaphore),
+        using_persistent_buffers(using_persistent_buffers),
+        chunks_per_sync(chunks_per_sync),
+        num_workers_per_link(num_workers_per_link),
+        num_buffers_per_channel(num_buffers_per_channel),
+        reverse_order(reverse_order),
+        sub_core_grid(sub_core_grid) {}
+
+    // Add attributes method for reflection
+    auto attributes() const {
+        using tt::stl::reflection::Attribute;
+        std::vector<std::tuple<std::string, Attribute>> attrs;
+
+        attrs.emplace_back("dim", dim);
+        attrs.emplace_back("num_links", num_links);
+        attrs.emplace_back("ring_size", ring_size);
+        attrs.emplace_back("output_mem_config", output_mem_config);
+        attrs.emplace_back("topology", topology);
+        attrs.emplace_back("barrier_semaphore", barrier_semaphore);
+        attrs.emplace_back("using_persistent_buffers", using_persistent_buffers);
+        attrs.emplace_back("cluster_axis", cluster_axis);
+        attrs.emplace_back("use_all_gather_async_llama_sharded", use_all_gather_async_llama_sharded);
+        attrs.emplace_back("use_optimal_ccl_for_llama", use_optimal_ccl_for_llama);
+        attrs.emplace_back("semaphore", semaphore);
+        attrs.emplace_back("chunks_per_sync", chunks_per_sync);
+        attrs.emplace_back("num_worker_per_link", num_workers_per_link);
+        attrs.emplace_back("num_buffers_per_channel", num_buffers_per_channel);
+        attrs.emplace_back("reverse_order", reverse_order);
+        attrs.emplace_back("sub_core_grid", sub_core_grid);
+        return attrs;
+    }
+
+    void validate_with_output_tensors(
+        const std::vector<Tensor>& input_tensors, const std::vector<std::optional<Tensor>>& output_tensors) const;
+    std::vector<ttnn::TensorSpec> compute_output_specs(const std::vector<Tensor>& input_tensors) const;
+    tt::tt_metal::operation::MeshWorkloadWithCallbacks create_mesh_workload(
+        const ttnn::MeshCoordinateRangeSet& tensor_coords,
+        const std::vector<Tensor>& input_tensors,
+        std::vector<Tensor>& output_tensors) const;
+    tt::tt_metal::operation::ProgramWithCallbacks create_program_at(
+        const ttnn::MeshCoordinate& coord,
+        const std::vector<Tensor>& input_tensors,
+        std::vector<Tensor>& output_tensors) const;
+    std::vector<Tensor> create_output_tensors(
+        const std::vector<Tensor>& input_tensors,
+        const std::vector<std::optional<Tensor>>& optional_output_tensors) const;
+    tt::tt_metal::operation::Hash compute_program_hash(const std::vector<Tensor>& input_tensors) const;
+
+    AllGatherAsyncVersion select_version(const Tensor& input_tensor) const;
+};
+
+struct AllGatherProgramArtifacts {
+    tt::tt_metal::KernelHandle reader_kernel_id;
+    tt::tt_metal::KernelHandle writer_kernel_id;
+    std::vector<tt::tt_metal::CoreCoord> all_cores;
+    uint32_t num_directions_per_link;
+    uint32_t num_workers_per_direction;
+    uint32_t num_mux_cores_per_direction_per_link;
+    uint32_t num_cores_per_link;
+};
+
+// Builder function that creates kernels and returns artifacts
+AllGatherProgramArtifacts build_all_gather_async_minimal_default_program_artifacts(
+    tt::tt_metal::Program& program,
+    const Tensor& input_tensor,
+    const MeshCoordinate& sender_device_coord,
+    const std::optional<MeshCoordinate>& forward_coord,
+    const std::optional<MeshCoordinate>& backward_coord,
+    Tensor& output_tensor,
+    uint32_t dim,
+    uint32_t num_links,
+    uint32_t ring_size,
+    uint32_t ring_index,
+    ccl::Topology topology,
+    const std::vector<GlobalSemaphore>& semaphore,
+    const std::optional<GlobalSemaphore>& barrier_semaphore,
+    bool using_persistent_buffers,
+    const std::optional<tt::tt_metal::SubDeviceId>& sub_device_id,
+    std::optional<experimental::ccl::AllGatherFusedOpSignaler>& fused_op_signaler,
+    std::optional<uint32_t> chunks_per_sync,
+    std::optional<uint32_t> num_workers_per_direction_opt,
+    std::optional<uint32_t> num_buffers_per_channel,
+    CoreCoord core_grid_offset,
+    bool reverse_order,
+    const std::optional<CoreRangeSet>& sub_core_grid = std::nullopt);
+
+// Runtime argument override function
+void all_gather_async_minimal_default_helper_override_runtime_arguments(
+    tt::tt_metal::Program& program,
+    tt::tt_metal::KernelHandle reader_kernel_id,
+    tt::tt_metal::KernelHandle writer_kernel_id,
+    const std::vector<tt::tt_metal::CoreCoord>& all_cores,
+    uint32_t num_links,
+    uint32_t num_directions_per_link,
+    uint32_t num_workers_per_direction,
+    uint32_t num_mux_cores_per_direction_per_link,
+    uint32_t num_cores_per_link,
+    const std::optional<GlobalSemaphore>& barrier_semaphore,
+    const std::vector<GlobalSemaphore>& semaphore,
+    const Tensor& input,
+    const Tensor& output);
+
+// All Gather Variants
+tt::tt_metal::operation::ProgramWithCallbacks all_gather_async_minimal_default(
+    const Tensor& input_tensor,
+    const MeshCoordinate& sender_device_coord,
+    const std::optional<MeshCoordinate>& forward_coord,
+    const std::optional<MeshCoordinate>& backward_coord,
+    Tensor& output_tensor,
+    uint32_t dim,
+    uint32_t num_links,
+    uint32_t ring_size,
+    uint32_t ring_index,
+    ccl::Topology topology,
+    const std::vector<GlobalSemaphore>& semaphore,
+    const std::optional<GlobalSemaphore>& barrier_semaphore,
+    bool using_persistent_buffers,
+    const std::optional<tt::tt_metal::SubDeviceId>& sub_device_id,
+    std::optional<uint32_t> chunks_per_sync,
+    std::optional<uint32_t> num_workers_per_link,
+    std::optional<uint32_t> num_buffers_per_channel,
+    bool reverse_order = false,
+    const std::optional<CoreRangeSet>& sub_core_grid = std::nullopt);
+
+tt::tt_metal::operation::ProgramWithCallbacks all_gather_async_minimal_default_helper(
+    tt::tt_metal::Program& program,
+    const Tensor& input_tensor,
+    const MeshCoordinate& sender_device_coord,
+    const std::optional<MeshCoordinate>& forward_coord,
+    const std::optional<MeshCoordinate>& backward_coord,
+    Tensor& output_tensor,
+    uint32_t dim,
+    uint32_t num_links,
+    uint32_t ring_size,
+    uint32_t ring_index,
+    ccl::Topology topology,
+    const std::vector<GlobalSemaphore>& semaphore,
+    const std::optional<GlobalSemaphore>& barrier_semaphore,
+    bool using_persistent_buffers,
+    const std::optional<tt::tt_metal::SubDeviceId>& sub_device_id,
+    std::optional<experimental::ccl::AllGatherFusedOpSignaler>& fused_op_signaler,
+    std::optional<uint32_t> chunks_per_sync,
+    std::optional<uint32_t> num_workers_per_link,
+    std::optional<uint32_t> num_buffers_per_channel,
+    CoreCoord core_grid_offset = CoreCoord(0, 0),
+    bool reverse_order = false,
+    const std::optional<CoreRangeSet>& sub_core_grid = std::nullopt);
+tt::tt_metal::operation::ProgramWithCallbacks all_gather_async_llama_sharded(
+    const Tensor& input_tensor,
+    const MeshCoordinate& sender_device_coord,
+    const std::optional<MeshCoordinate>& forward_coord,
+    const std::optional<MeshCoordinate>& backward_coord,
+    Tensor& output_tensor,
+    uint32_t dim,
+    uint32_t num_links,
+    uint32_t ring_size,
+    uint32_t ring_index,
+    ccl::Topology topology,
+    const GlobalSemaphore& semaphore,
+    const std::optional<GlobalSemaphore>& barrier_semaphore,
+    bool using_persistent_buffers,
+    const std::optional<tt::tt_metal::SubDeviceId>& sub_device_id,
+    bool use_optimal_ccl_for_llama);
+
+namespace operations::experimental::ccl {
+
+Tensor all_gather_async(
+    const Tensor& input_tensor,
+    uint32_t dim,
+    const std::vector<GlobalSemaphore>& multi_device_global_semaphore,
+    uint32_t num_links = 1,
+    const std::optional<MemoryConfig>& memory_config = std::nullopt,
+    ttnn::ccl::Topology topology = ttnn::ccl::Topology::Ring,
+    std::optional<tt::tt_metal::SubDeviceId> sub_device_id = std::nullopt,
+    bool use_optimal_ccl_for_llama = false,
+    bool use_all_gather_async_llama_sharded = false,
+    const std::optional<GlobalSemaphore>& barrier_semaphore = std::nullopt,
+    bool reverse_order = false,
+    const std::optional<CoreRangeSet>& sub_core_grid = std::nullopt);
+
+Tensor all_gather_async(
+    const Tensor& input_tensor,
+    const std::optional<ttnn::Tensor>& persistent_output_buffer,
+    uint32_t dim,
+    const std::vector<GlobalSemaphore>& multi_device_global_semaphore,
+    uint32_t num_links = 1,
+    const std::optional<MemoryConfig>& memory_config = std::nullopt,
+    ttnn::ccl::Topology topology = ttnn::ccl::Topology::Ring,
+    std::optional<tt::tt_metal::SubDeviceId> sub_device_id = std::nullopt,
+    std::optional<uint32_t> cluster_axis = std::nullopt,
+    bool use_optimal_ccl_for_llama = false,
+    bool use_all_gather_async_llama_sharded = false,
+    const std::optional<GlobalSemaphore>& barrier_semaphore = std::nullopt,
+    std::optional<uint32_t> chunks_per_sync = std::nullopt,
+    std::optional<uint32_t> num_workers_per_link = std::nullopt,
+    std::optional<uint32_t> num_buffers_per_channel = std::nullopt,
+    bool reverse_order = false,
+    const std::optional<CoreRangeSet>& sub_core_grid = std::nullopt);
+
+// same as above but for vector of mesh
+std::vector<Tensor> all_gather_async(
+    const std::vector<Tensor>& input_tensors,
+    const std::optional<ttnn::Tensor>& persistent_output_buffer,
+    uint32_t dim,
+    const std::vector<global_semaphore::MultiDeviceGlobalSemaphore>& multi_device_global_semaphore,
+    uint32_t num_links = 1,
+    const std::optional<MemoryConfig>& memory_config = std::nullopt,
+    ttnn::ccl::Topology topology = ttnn::ccl::Topology::Ring,
+    std::optional<tt::tt_metal::SubDeviceId> sub_device_id = std::nullopt,
+    std::optional<uint32_t> cluster_axis = std::nullopt,
+    bool use_all_gather_async_llama_sharded = false,
+    bool use_optimal_ccl_for_llama = false,
+    const std::optional<std::vector<GlobalSemaphore>>& barrier_semaphore = std::nullopt,
+    std::optional<uint32_t> chunks_per_sync = std::nullopt,
+    std::optional<uint32_t> num_workers_per_link = std::nullopt,
+    std::optional<uint32_t> num_buffers_per_channel = std::nullopt,
+    bool reverse_order = false,
+    const std::optional<CoreRangeSet>& sub_core_grid = std::nullopt);
+
+Tensor all_gather_async(
+    const Tensor& input_tensor,
+    int32_t dim,
+    uint32_t cluster_axis,
+    const MeshDevice& mesh_device,
+    ttnn::ccl::Topology topology,
+    const std::vector<GlobalSemaphore>& multi_device_global_semaphore,
+    const std::optional<ttnn::Tensor>& persistent_output_tensor = std::nullopt,
+    const std::optional<MemoryConfig>& memory_config = std::nullopt,
+    std::optional<size_t> num_preferred_links = std::nullopt,
+    std::optional<tt::tt_metal::SubDeviceId> sub_device_id = std::nullopt,
+    bool use_all_gather_async_llama_sharded = false,
+    bool use_optimal_ccl_for_llama = false,
+    const std::optional<GlobalSemaphore>& barrier_semaphore = std::nullopt,
+    bool reverse_order = false,
+    const std::optional<CoreRangeSet>& sub_core_grid = std::nullopt);
+
+}  // namespace operations::experimental::ccl
+
+}  // namespace ttnn
