+++ ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_matmul_async/device/all_gather_matmul_async_program_factory.hpp
@@ -14,11 +14,10 @@
 namespace ttnn::operations::experimental::ccl::all_gather_matmul_async::program {
 
 struct AllGatherMatmulAsyncSharedVariables {  // TODO: migrate shared_variables to not use the old
-                                              // override_runtime_arguments_callback in matmul old
+                                              // override_runtime_arguments_callback in matmul and all_gather_async old
                                               // program factories
     std::optional<tt::tt_metal::operation::OverrideRuntimeArgumentsCallback<std::vector<Tensor>>>
         all_gather_override_runtime_arguments_callback;
-    all_gather_async::AllGatherProgramArtifacts all_gather_async_shared_variables;
     std::optional<tt::tt_metal::operation::OverrideRuntimeArgumentsCallback<Tensors>>
         matmul_override_runtime_arguments_callback;
 };
