+++ ttnn/cpp/ttnn/operations/experimental/ccl/composite_common.hpp
@@ -5,7 +5,7 @@
 
 #pragma once
 
-#include "ttnn/operations/experimental/ccl/all_gather_async/device/all_gather_async_device_operation.hpp"
+#include "ttnn/operations/experimental/ccl/all_gather_async/device/all_gather_async_op.hpp"
 #include "ttnn/operations/ccl/all_broadcast/device/all_broadcast_device_operation.hpp"
 #include "ttnn/operations/experimental/ccl/reduce_scatter_minimal_async/device/reduce_scatter_minimal_async_op_device_operation.hpp"
 #include "ttnn/types.hpp"
